comment delete view

+++ app/controllers/comments_controller.rb
@@ -4,8 +4,7 @@ class CommentsController < ApplicationController
 def create 
 	@comment = current_user.comments.build(comment_params)
 	if @comment.save
-			flash.now[:notice] = "Comment posted!!! "
-
+			flash[:notice] = "Comment posted!!! "
 		redirect_to article_path(@comment.article.id)
 	else
 		flash[:error] = "Something went wrong"
@@ -14,12 +13,17 @@ end
 
 def destroy
 	@comment = Comment.find(params[:id])
-	if @comment.destroy
-				flash[:error] = "Comment deleted!!!"
-		redirect_to articles_path
+	@article = @comment.article
+	if @comment.user.id == current_user.id
+		if @comment.destroy
+			flash[:error] = "Comment deleted!!!"
+			redirect_to @article
+		else
+			flash[:error] = "Something went wrong"
+			redirect_to articles_path
+		end
 	else
-		flash[:error] = "Something went wrong"
-		redirect_to articles_path
+		flash[:notice] = "Can only delete your comment"
 	end
 end
